pyproject.toml
[project]
name = "surveyagent"
version = "0.1.0"
description = "AI-powered conversational survey platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "motor>=3.7.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "openai>=1.60.0",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.1",
    "email-validator>=2.0.0",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
]
